Add a snapshot copy of SORAL
This is a single snapshot of SORAL, combined from a checkout of the
subversion repository and a slightly older checkout of the CVS repository.

I did this because I found that the autoconfiscated version of the
build system from the SVN checkout I had didn't actually work on my
system, some oddness of how the autoconf was set up.  Rather than
futz with it, I just rolled back to the old pre-autoconf CVS version I
had, which did not differ in any way in the main source code from the
svn version, but built fine on my system.