Remove GNU make specific ifeq construct
ifeq/else is a GNU makeism, and doesn't work with BSD make.  Since the
whole point of the ifeq was to emit a "we don't know how to work on
Windows" message, let's just leave it out.

Also, the default target built "libSoral.a" but was labeled
"libsoral.a".  Let's make the target actually have the same name as
the thing it builds, so that repeated makes don't waste effort.